Further refactor replica code
