+++ replica-node/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod replica_node;
+pub mod replica_util;
